+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Enhancements:
   - Adds methods to CILviewer and CILviewer2D #425
   
 Bugfix:
+  - Differentiate 3D and 2D images in the converter `numpy2vtkImage` #437
   - Edit slider min value #420
   - Fix extent error when user clicks in the viewer to create a box by clicking outside of the image #425
   - Fix extent error when user enlarges the box outside the image #425